Upgrade Postgres to 9.2, add dependencies
The apt cookbook required a more recent version of Chef to work,
so I added an initial provisioning that upgrades Chef before
the rest of the recipes run. Chef 11 is also pickier about
dependencies, so I added a few more cookbooks.

Postgres is now from a different ppa --one that won't be deprecated
in the near future. It is also now using the generic database
cookbook for generating tables and users.