Add option of using omnibus_updater. Clean config.
I still use a custom recipe instead of the omnibus
updater to install the updated chef gem
since, for this application, the whole kit'n'kaboodle isn't
needed, and it's faster.

I also removed much of the boilerplate documentation in the
Vagrantfile to de-clutter it and make it more straightforward
for this particular purpose (e.g. we don't need puppet config
options in there).